+++ src/App.tsx
@@ -2,25 +2,13 @@ import { useState } from 'react'
 import Logo from './assets/doctor-svgrepo-com.svg'
 import clockLogo from './assets/clock-circle-svgrepo-com.svg'
 import './App.css'
-import Button from './Button';
-import { device } from './styles/BreakPoints'
-import { theme } from './styles/Theme'
-import styled, { ThemeProvider } from 'styled-components'
-
-const Title = styled.h1`
-  color: ${(props) => props.theme.colors.primaryTextColor};
-  font-size: 48px;
-  @media ${device.md} {
-    font-size: 32px;
-  }
-`
+import Button from './components/Button';
 
 
 function App() {
   const [count, setCount] = useState(0)
 
   return (
-    <ThemeProvider theme={theme}>
     <>
       <div>
         <a href="" target="_blank">
@@ -46,7 +34,6 @@ function App() {
         allskonar frábært
       </p>
     </>
-    </ThemeProvider>
   )
 }
